Altered comment testing
factored check_comment out of check_comments and adopted the format of
comments read from json input with explict value and name names in the
list representation. There was previously some inconsistency with list
key value pairs being mapped to name and value as opposed to the
approach that is now to be use everywhere. There were some comments with
duplicated names which is incompatible with the previous approach.